+++ core/mdate.h
@@ -40,6 +40,7 @@ public:
     Q_INVOKABLE static QString formatDate(const QDate &date, int currentYear, const QString &type = "");
     Q_INVOKABLE static int season(const QDate &date);
     Q_INVOKABLE static int seasonYear(const QDate &date);
+    Q_INVOKABLE static qint64 daysTo(const QDate &from, const QDate &to) { return from.daysTo(to); }
 };
 
 #endif // MDATE_H